modules/SampleModule/requirements.txt
azure-iot-device~=2.9.0
pyModbusTCP==0.1.10
